+++ src/Commands/1_music/save.js
@@ -2,6 +2,7 @@ const Command = require('../../Structures/Command');
 
 const queue = require('../../Data/queue');
 const { emoji: { success, warning }, response: { noMusic } } = require('../../../config/config.json');
+const { consoleLog } = require('../../Data/Log');
 
 
 module.exports = new Command({
@@ -12,7 +13,13 @@ module.exports = new Command({
 		const guildQueue = queue.get(message.guild.id);
         if (!guildQueue) return message.channel.send(`${warning} ${noMusic}`);
 
-		message.channel.send(`${success} Sending the song to you!`);
-		message.author.send(`${success} You saved the track **${guildQueue.songs[0].title}** (${guildQueue.songs[0].url}) from the server **${message.guild.name}**`);
+		try {
+			message.author.send(`${success} You saved the track **${guildQueue.songs[0].title}** (${guildQueue.songs[0].url}) from the server **${message.guild.name}**`);
+			message.channel.send(`${success} I sent the song to your DMs!`);
+		}
+		catch (error) {
+			consoleLog('[WARN] Could not send DM to user', error);
+			message.channel.send(`${warning} I couldn't send you a DM.`);
+		}
 	}
 });