+++ src/Commands/1_music/shuffle.js
@@ -1,7 +1,7 @@
 const Command = require('../../Structures/Command');
 
 const queue = require('../../Data/queue');
-const { emoji: { success, warning }, response: { wrongChannel, noMusic } } = require('../../../config/config.json');
+const { bot: { ownerID }, emoji: { success, warning }, response: { wrongChannel, noMusic } } = require('../../../config/config.json');
 
 
 module.exports = new Command({
@@ -12,9 +12,9 @@ module.exports = new Command({
 		const guildQueue = queue.get(message.guild.id);
         if (!guildQueue) return message.channel.send(`${warning} ${noMusic}`);
 
-        if (!message.member.voice.channel || guildQueue.connection.joinConfig.channelId != message.member.voice.channel.id) return message.channel.send(`${warning} ${wrongChannel}`);
+        if ((!message.member.voice.channel || guildQueue.connection.joinConfig.channelId != message.member.voice.channel.id) && message.author.id != ownerID) return message.channel.send(`${warning} ${wrongChannel}`);
 
 		queue.shuffle(message.guild.id);
-		message.channel.send(`${success} You successfully shuffled your ${guildQueue.songs.length - 1} entries.`);
+		message.channel.send(`${success} You successfully shuffled ${guildQueue.songs.length - 1} entries.`);
 	}
 });