+++ src/Commands/3_owner/setgame.js
@@ -1,5 +1,7 @@
 const Command = require('../../Structures/Command');
 
+const { ActivityType } = require('discord.js');
+
 const { bot: { ownerID }, status: { game }, emoji: { success, error }, response: { invalidPermissions } } = require('../../../config/config.json');
 
 
@@ -13,13 +15,13 @@ module.exports = new Command({
         if (!args[0]) {
 			client.user.setActivity({
 				name: game,
-				type: 'PLAYING'
+				type: ActivityType.Playing
 			});
 			return message.channel.send(`${success} I'm now playing \`**${game}**\` (default)`);
 		}
         client.user.setActivity({
 			name: args.slice(0).join(' '),
-			type: 'PLAYING'
+			type: ActivityType.Playing
 		});
         message.channel.send(`${success} I'm now playing \`${args.slice(0).join(' ')}\``);
 	}